pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uni-d2"
version = "0.1.0"
description = "A modular framework for discrete diffusion models including MDLM, SEDD, UDLM, BD3LM, GIDD, and more"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Kalyan Nadimpalli"}
]
keywords = ["diffusion", "discrete-diffusion", "machine-learning", "deep-learning", "pytorch", "lightning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Generic
    "hydra-core==1.3.2",
    "tqdm==4.67.1",
    "termcolor==3.0.1",
    "seaborn==0.13.2",
    "blobfile==3.0.0",
    "fsspec",
    "git-lfs==1.6",
    "jinja2==3.1.5",
    "pyyaml==6.0.2",
    "torchmetrics==1.7.1",
    
    # ML
    "torch==2.7.0",
    "einops==0.8.1",
    "fancy-einsum==0.0.3",
    "numpy==1.26.4",
    "mauve-text==0.4.0",
    "sentencepiece==0.2.0",
    "lightning==2.5.1",
    "pandas==2.2.1",
    "scikit-learn==1.4.0",
    "timm==1.0.15",
    "jupyter==1.1.1",
    "ocifs==1.3.2",
    
    # Huggingface
    "huggingface-hub==0.30.2",
    "safetensors==0.5.2",
    "tokenizers==0.20.3",
    "evaluate==0.4.3",
    "datasets==3.5.0",
    "transformers==4.45.0",
    "accelerate==1.6.0",
    
    # Monitoring / configuration
    "tensorboard==2.16.2",
    "wandb==0.19.9",
    "omegaconf==2.3.0",
    "rich==13.9.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mkdocs",
    "mkdocs-material",
    "mkdocstrings[python]",
    "mkdocs-include-markdown-plugin",
]
flash-attn = [
    "flash-attn>=2.7.0",
]

[project.scripts]
uni-d2 = "discrete_diffusion.__main__:main"

[project.urls]
Homepage = "https://github.com/nkalyanv/uni-d2"
Documentation = "https://nkalyanv.github.io/uni-d2/"
Repository = "https://github.com/nkalyanv/uni-d2"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["discrete_diffusion*"]
namespaces = false


